It usually works better with a double to handle numbers between 0 and 1

+++ blend/solver.h
@@ -499,7 +499,7 @@ pastix_int_t solverSave(const SolverMatrix *solvptr, FILE *stream);
 
 int solverComputeGPUDistrib( SolverMatrix *solvmtx,
                              int           ngpus,
-                             int           memory_percentage,
+                             double        memory_percentage,
                              size_t        eltsize,
                              int           criterium,
                              int           factotype );
